fastapi/pyproject.toml
[project]
name = "dv-admin-fastapi"
version = "0.1.0"
description = "DV-Admin FastAPI Backend"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "DV-Admin Team"}
]
keywords = ["fastapi", "tortoise-orm", "admin", "async"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # FastAPI 核心
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    # 数据库 - 同时支持 SQLite 和 MySQL
    "tortoise-orm[asyncpg]>=0.24.0",
    "asyncmy>=0.2.0",
    "aiomysql>=0.2.0",
    "aiosqlite>=0.20.0", # SQLite 异步支持
    # Redis 缓存
    "redis>=5.0.0",
    # 认证与安全
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "bcrypt==3.2.2",
    # 配置与验证
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    # 工具
    "python-dotenv>=1.0.0",
    "email-validator>=2.2.0",
    "orjson>=3.10.0",
    # 日志
    "loguru>=0.7.0",
    # 验证码
    "captcha>=0.4.0",
    # Excel 处理
    "openpyxl>=3.1.0",
]

[project.scripts]
dv-admin = "app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*"]

[tool.uv]
package = true

# Black 代码格式化配置
[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort 导入排序配置
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
known_first_party = ["app"]

# Ruff 代码检查配置
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
]
ignore = [
    "E501",  # line too long, handled by black
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["E501", "F401", "F841"]
"app/db/import_django_data.py" = ["E402"]
"app/db/verify_import.py" = ["E402"]

# MyPy 类型检查配置
[tool.mypy]
python_version = "3.10"
check_untyped_defs = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_decorators = false
no_implicit_optional = false
warn_redundant_casts = true
warn_unused_ignores = false
warn_return_any = false
strict_equality = false
show_error_codes = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Pytest 测试配置
[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# Coverage 覆盖率配置
[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
fail_under = 80

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.6.0",
]
